video_copy/requirements.txt
granian
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.6.0; python_version >= '3.9'
asttokens==2.4.1
attrs==24.2.0; python_version >= '3.7'
certifi==2024.8.30; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
comm==0.2.2; python_version >= '3.8'
confluent-kafka==2.5.3
decorator==5.1.1; python_version >= '3.5'
executing==2.1.0; python_version >= '3.8'
fastapi==0.115.0; python_version >= '3.8'
fastjsonschema==2.20.0
filelock==3.16.1; python_version >= '3.8'
fsspec==2024.9.0; python_version >= '3.8'
grpcio==1.66.2; python_version >= '3.8'
grpcio-tools==1.66.2; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
h2==4.1.0
hpack==4.0.0; python_full_version >= '3.6.1'
httpcore==1.0.5; python_version >= '3.8'
httpx[http2]==0.27.2; python_version >= '3.8'
huggingface-hub==0.25.1; python_full_version >= '3.8.0'
hyperframe==6.0.1; python_full_version >= '3.6.1'
idna==3.10; python_version >= '3.6'
ipython==8.27.0; python_version >= '3.10'
ipywidgets==8.1.5; python_version >= '3.7'
jedi==0.19.1; python_version >= '3.6'
jinja2==3.1.4; python_version >= '3.7'
joblib==1.4.2; python_version >= '3.8'
jsonschema==4.23.0; python_version >= '3.8'
jsonschema-specifications==2023.12.1; python_version >= '3.8'
jupyter-core==5.7.2; python_version >= '3.8'
jupyterlab-widgets==3.0.13; python_version >= '3.7'
markupsafe==2.1.5; python_version >= '3.7'
matplotlib-inline==0.1.7; python_version >= '3.8'
mpmath==1.3.0
nbformat==5.10.4; python_version >= '3.8'
networkx==3.3; python_version >= '3.10'
numpy==2.1.1; python_version >= '3.10'
nvidia-cublas-cu12==12.1.3.1; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cuda-cupti-cu12==12.1.105; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cuda-nvrtc-cu12==12.1.105; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cuda-runtime-cu12==12.1.105; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cudnn-cu12==9.1.0.70; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cufft-cu12==11.0.2.54; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-curand-cu12==10.3.2.106; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cusolver-cu12==11.4.5.107; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cusparse-cu12==12.1.0.106; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-nccl-cu12==2.20.5; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-nvjitlink-cu12==12.6.68; python_version >= '3'
nvidia-nvtx-cu12==12.1.105; platform_system == 'Linux' and platform_machine == 'x86_64'
packaging==24.1; python_version >= '3.8'
pandas==2.2.3; python_version >= '3.9'
parso==0.8.4; python_version >= '3.6'
pexpect==4.9.0; sys_platform != 'win32' and sys_platform != 'emscripten'
pillow==10.4.0; python_version >= '3.8'
platformdirs==4.3.6; python_version >= '3.8'
plotly==5.24.1; python_version >= '3.8'
polars==1.8.2; python_version >= '3.8'
portalocker==2.10.1; python_version >= '3.8'
prompt-toolkit==3.0.48; python_full_version >= '3.7.0'
protobuf==5.28.2; python_version >= '3.8'
ptyprocess==0.7.0
pure-eval==0.2.3
pydantic==2.9.2; python_version >= '3.8'
pydantic-core==2.23.4; python_version >= '3.8'
pydantic-settings==2.5.2; python_version >= '3.8'
pygments==2.18.0; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2024.2
pyyaml==6.0.2; python_version >= '3.8'
qdrant-client==1.11.3; python_version >= '3.8'
referencing==0.35.1; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
rpds-py==0.20.0; python_version >= '3.8'
safetensors==0.4.5; python_version >= '3.7'
scikit-learn==1.5.2; python_version >= '3.9'
scipy==1.14.1; python_version >= '3.10'
setuptools==75.1.0; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
stack-data==0.6.3
starlette==0.38.6; python_version >= '3.8'
sympy==1.13.3; python_version >= '3.8'
tenacity==9.0.0; python_version >= '3.8'
threadpoolctl==3.5.0; python_version >= '3.8'
timm==1.0.9; python_version >= '3.8'
torch==2.4.1; python_full_version >= '3.8.0'
torchvision==0.19.1; python_version >= '3.8'
tqdm==4.66.5; python_version >= '3.7'
traitlets==5.14.3; python_version >= '3.8'
triton==3.0.0; python_version < '3.13' and platform_system == 'Linux' and platform_machine == 'x86_64'
typing-extensions==4.12.2; python_version >= '3.8'
tzdata==2024.2; python_version >= '2'
urllib3==2.2.3; python_version >= '3.8'
uvicorn==0.31.0; python_version >= '3.8'
wcwidth==0.2.13
widgetsnbextension==4.0.13; python_version >= '3.7'